eatgo-customer-api/src/main/java/com/example/eatgo/service/MenuItemService.java
package com.example.eatgo.service;

import com.example.eatgo.domain.MenuItem;
import com.example.eatgo.repository.MenuItemRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
@RequiredArgsConstructor
public class MenuItemService {
    private final MenuItemRepository menuItemRepository;

    public void  bulkUpdate(Long restaurantId, List<MenuItem> menuItems) {
        for(MenuItem menuItem : menuItems){
            updateOrDeleteMenuItems(restaurantId, menuItem);
        }
    }

    private void updateOrDeleteMenuItems(Long restaurantId, MenuItem menuItem) {
        if(menuItem.isDestroy()){
            menuItemRepository.deleteById(menuItem.getId());
        }else {
            menuItem.setRestaurantId(restaurantId);
            menuItemRepository.save(menuItem);
        }
    }
}
